+++ base.ts
@@ -1,8 +1,5 @@
-import globalAxios, { AxiosInstance } from "axios";
-import * as http from "http";
-import * as https from "https";
-
 import { Configuration, UserConfigurationParams } from "./configuration";
+import { HttpClient } from "./common";
 import { Credentials } from "./credentials";
 
 const DEFAULT_CONNECTION_TIMEOUT_IN_MS = 10000;
@@ -25,8 +22,9 @@ export interface RequestArgs {
 export class BaseAPI {
   protected configuration: Configuration;
   protected credentials: Credentials;
+  protected httpClient: HttpClient;
 
-  constructor(configuration: UserConfigurationParams | Configuration, protected axios?: AxiosInstance) {
+  constructor(configuration: UserConfigurationParams | Configuration, httpClient?: HttpClient) {
     if (configuration instanceof Configuration) {
       this.configuration = configuration;
     } else {
@@ -34,17 +32,12 @@ export class BaseAPI {
     }
     this.configuration.isValid();
 
-    this.credentials = Credentials.init(this.configuration, this.axios);
+    this.httpClient = httpClient ?? {
+      fetch: globalThis.fetch.bind(globalThis),
+      defaultTimeout: DEFAULT_CONNECTION_TIMEOUT_IN_MS,
+      defaultHeaders: this.configuration.baseOptions?.headers,
+    };
 
-    if (!this.axios) {
-      const httpAgent = new http.Agent({ keepAlive: true });
-      const httpsAgent = new https.Agent({ keepAlive: true });
-      this.axios = globalAxios.create({
-        httpAgent,
-        httpsAgent,
-        timeout: DEFAULT_CONNECTION_TIMEOUT_IN_MS,
-        headers: this.configuration.baseOptions?.headers,
-      });
-    }
+    this.credentials = Credentials.init(this.configuration, this.httpClient);
   }
 }